requirements.txt
async-timeout
discord
Pillow
DiscordUtils
gspread